+++ pkg/agent/collectors/metrics/configmap_test.go
@@ -13,39 +13,27 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clusterv1 "sigs.k8s.io/cluster-api/api/core/v1beta1"
 
-	"github.com/giantswarm/observability-operator/pkg/auth"
 	"github.com/giantswarm/observability-operator/pkg/common/organization/mocks"
 	"github.com/giantswarm/observability-operator/pkg/config"
+	"github.com/giantswarm/observability-operator/pkg/credential"
 	"github.com/giantswarm/observability-operator/pkg/domain/organization"
 )
 
 var managementClusterName = "dummy-cluster"
 
-// mockAuthManager implements auth.AuthManager for testing
-type mockAuthManager struct{}
+// mockCredentialReader implements credential.Reader for testing
+type mockCredentialReader struct{}
 
-func (m *mockAuthManager) EnsureClusterAuth(ctx context.Context, cluster *clusterv1.Cluster) error {
-	return nil
+func (m *mockCredentialReader) ReadPassword(ctx context.Context, namespace, credentialName string) (string, string, error) {
+	return "test-user", "test-password", nil
 }
 
-func (m *mockAuthManager) DeleteClusterAuth(ctx context.Context, cluster *clusterv1.Cluster) error {
-	return nil
-}
-
-func (m *mockAuthManager) GetClusterPassword(ctx context.Context, cluster *clusterv1.Cluster) (string, error) {
-	return "test-password", nil
-}
-
-func (m *mockAuthManager) DeleteGatewaySecrets(ctx context.Context) error {
-	return nil
-}
-
-var _ auth.AuthManager = &mockAuthManager{}
+var _ credential.Reader = &mockCredentialReader{}
 
 func newTestService(monitoringEnabled, exemplarsEnabled bool) *Service {
 	return &Service{
 		OrganizationRepository: mocks.NewMockOrganizationRepository("dummy-org"),
-		AuthManager:            &mockAuthManager{},
+		CredentialReader:       &mockCredentialReader{},
 		Config: config.Config{
 			Cluster: config.ClusterConfig{
 				BaseDomain: "test.gigantic.io",